added isbn lookup, replaced some refs that caused errors, added citation generation check for refs upon model creation, moved references_to_markdown out of web utils to core utils so that core can use that function without importing the builtins
