+++ requirements.txt
@@ -6,6 +6,7 @@ dash-table-experiments>=0.6.0
 frozendict==1.2
 gunicorn>=19.7.1
 habanero==0.6.0
+isbnlib>=3.9.6
 monty==1.0.3
 networkx>=2.0
 numpy>=1.15.1
